requirements.txt
mcp
python-dotenv
pydantic
sqlalchemy
pymongo
